Refactor out all of the switch nonsense
