+++ src/clippednoteentry.ts
